Trekker ut tekstinnhold til egen fil
